+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.17.0] - 2017-02-04
+
 ### Added
 
 - Test runner interface for test runners (already implemented by Godog and Main test runners)
@@ -282,7 +285,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/goph/fxt/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/goph/fxt/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/goph/fxt/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/goph/fxt/compare/v0.14.0...v0.15.0